pyproject.toml
[project]
name = "text-to-image-pipeline"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "accelerate",
    "diffusers @ git+https://github.com/huggingface/diffusers.git",
    "gradio",
    "python-dotenv",
    "sentencepiece",
    "torch",
    "transformers",
]

[dependency-groups]
dev = [
    "pytest",
    "ruff",
    "ty",
]
